requirements.txt
python-dotenv~=0.21.0
beautifulsoup4~=4.11.1
requests~=2.28.1
lxml~=4.9.1